requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cssselect==1.0.3
decorator==4.2.1
Flask==0.12.2
idna==2.6
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jupyter-client==5.2.2
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
nltk==3.2.5
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pyzmq==16.0.4
readability-lxml==0.6.2
reportlab==3.4.0
requests==2.18.4
selenium==3.9.0
simplegeneric==0.8.1
six==1.11.0
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
